requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
django-autoslug==1.9.8
django-tinymce==3.6.1
filelock==3.9.0
Pillow==9.4.0
platformdirs==3.1.0
sqlparse==0.4.3
virtualenv==20.20.0
